Ignoring assets with .import and .log extensions by introducing an asset loader matching these extensions that does not poll the async reader, hence not reading the file bytes. It instead simply inserts a unit type as the loaded asset.
